+++ modules/expo-bluesky-swiss-army/android/src/main/java/expo/modules/blueskyswissarmy/referrer/ExpoBlueskyReferrerModule.kt
@@ -1,4 +1,4 @@
-package expo.modules.blueskyreferrer
+package expo.modules.blueskyswissarmy.referrer
 
 import android.util.Log
 import com.android.installreferrer.api.InstallReferrerClient
